requirements.txt
autopep8==1.5.7
certifi==2021.5.30
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy @ file:///D:/bld/numpy_1626648564420/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.3.56
Pillow @ file:///C:/ci/pillow_1625663294916/work
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
wincertstore==0.2
